Fix pattern comparison on windows
This commit uses filepath.ToSlash to ensure all paths use forward seperators if on Windows.

The example `header_ignore` paths all use unix line endings which work fine so long as the paths compared are using unix path seperators. If this is run on Windows, filepath.Walk returns paths with Windows path seperators, which will appear different to doublestar.Match. It's safe to convert \ to / in place in this scenario, as these are valid paths.