hbs: update comment to remove line break
Template languages that generate HTML files, like Handlebars, are often
sensitive to whitespace, so adding a line break may change how a page is
rendered.

This commit adds a `~` at the end of the comment as a whitespace control
so the rendering engine removes the additional line break added by the
license header comment.