fix(gha): output headers cmd to stdout, not stderr (#80)
Previously, the GitHub Actions `::startgroup::` and `::endgroup` directives were being sent to stdout, but all command output from Copywrite was being sent to stderr. This caused an issue that was especially common for the `copywrite headers --plan` command, where the list of files missing headers would be partially included and partially excluded from the logging group.

As a fix I set Cobra up to print to stdout by default, and updated the logger we send to the addlicense module to also use stdout. Ideally, we'd restructure logging entirely within the module, but I'm leaving that for a future PR.