Ignore node_modules subdirectories (#109)
This adjusts the default ignore pattern for node_modules to include subdirectories when checking for headers. This is necessary because the node_modules directory is sometimes nested within other directories, and the current pattern only matches the top-level node_modules directory.

This really only affects interactive use, as node_modules directories are not usally committed to version control.