fix(pre-commit): Hooks should use the copywrite binary
If you use copywrite in your repo, you probably already have copywrite
installed.

The hooks running `go run .` doesn't work in my terraform provider,
because it is a plugin:

```
This binary is a plugin. These are not meant to be executed directly.
Please execute the program that consumes these plugins, which will
load any plugins automatically
exit status 1
```

So switch the entry point to be the copywrite program.
Don't pass filenames. copywrite knows what to do.

Pass arguments directly on the entry.

Closes #78

Signed-off-by: Bryan Hundven <bryanhundven@gmail.com>